Add files via upload
This was the cleaning and aggregation of raw price and volume data on bitcoin to see if it showed continuation on the daily chart similar to the popular 3 bar technical strategy. The results were not statistically significant with p=.1748. The parameters I chose were the first day being a price increase greater than 5% and volume being 50% greater than the 30 day moving average.